Calculate and add \(\alpha\)

Calculate \(\alpha\) for a two source trophic position model using equations from Post 2002.

Usage

add_alpha(data)

Arguments

data

data.frame of stable isotope samples with mean values for two baselines. For aquatic ecosystems, baseline two needs to come from a pelagic source, while baseline one needs to come from a benthic source. Baseline \(\delta^1^3 Cc\) columns need to be named c1 and c2, with the consumer's column named d13c.

Value

a data.frame that has alpha, min_alpha, and max_alpha added.

Details

$$\alpha = (\delta^1^3 Cc - \delta ^1^3Cp) / (\delta ^1^3Cb - \delta ^1^3Cb)$$

where \(\delta^1^3 Cc\) is the isotopic value for consumer, \(\delta^1^3 Cp\) is the mean isotopic value for pelagic baseline, and \(\delta^1^3 Cb\) is the mean isotopic value for benthic baseline.

Examples

example_iso_tsn |>
  add_alpha()
#> Error: object 'example_iso_tsn' not found