trps: Bayesian trophic position models using stan

Bayesian trophic position models using stan by leveraging 'brms' for stable isotope data. Trophic position models are derived by using equations from Post (2002) doi:10.1890/0012-9658(2002)083[0703:USITET]2.0.CO;2 , and Huevel et al. (2024) doi:10.1139/cjfas-2024-0028 .

See also

Author

Maintainer: Benjamin L. Hlina benjamin.hlina@gmail.com