Bayesian model - Two Source Trophic Position with \(\alpha_c\) and carbon source

Bayesian model for trophic position using a two source model with \(\alpha_c\) and carbon source derived from Post 2002 and Heuvel et al. 2024

Usage

two_source_model_acc(bp = FALSE, lambda = NULL)

Arguments

bp

logical value that controls whether priors are supplied to the model for \(\delta\)\(^1\)\(^5\)N pelagic and benthic baselines. Default is FALSE meaning the model will not be supplied priors, however, the supplied data.frame needs values for \(\delta\)\(^1\)\(^5\)N each baseline (n1 and n2).

lambda

numerical value, 1 or 2, that controls whether one or two lambdas are used. See details for equations and when to use 1 or 2. Defaults to 1.

Value

returns model structure for two source model to be used in a brms() call.

See also

Examples

two_source_model_acc()
#> alpha ~ ac * (max_alpha - min_alpha) + min_alpha 
#> ac ~ 1
#> d13c ~ ac * (c1 - c2) + c2 
#> ac ~ 1
#> d15n ~ dn * (tp - l) + n1 * ac + n2 * (1 - ac) 
#> ac ~ 1
#> tp ~ 1
#> dn ~ 1