Bayesian priors - Two Source Trophic Position

Create priors for two source trophic position model derived from Post 2002.

Usage

two_source_priors(bp = FALSE)

Arguments

bp

logical value that controls whether informed priors are supplied to the model for both \(\delta^{15}\)N baselines. Default is FALSE meaning the model will use uninformed priors, however, the supplied data.frame needs values for both \(\delta^{15}\)N baseline (n1 and n2).

Value

returns priors for two source model to be used in a brms() call.

See also

Examples

two_source_priors()
#>                        prior class      coef group resp dpar nlpar    lb
#>                   beta(a, b)     b                           alpha     0
#>         normal(dn, dn_sigma)     b Intercept                    dn  <NA>
#>        uniform(tp_lb, tp_ub)     b                              tp tp_lb
#>  uniform(sigma_lb, sigma_ub) sigma                                  <NA>
#>        ub source
#>         1   user
#>      <NA>   user
#>     tp_ub   user
#>  sigma_ub   user